+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Les changements notables sont documentés dans ce fichier. Ce projet respecte le
 
 ### Ajouts
 - Ajout de type de grilles supplémentaires :
-	- Grille en losage
+	- Grille en losange
 	- Grille rectangulaire
 
 ## Version 1.4 [15/05/2016]